refactor(handler): move pgtype-to-pointer helpers to pgptr.go (#1022)
Pure code motion: the eight read-side pgtype-to-pointer helpers
(textPtrValue, uuidPtrValue, timePtr, intPtrValue, float4PtrValue,
int8PtrValue, boolPtrValue, int32PtrValue) move verbatim from workers.go
into the new pgptr.go. --color-moved clean; no exported-API change and no
behaviour change (go doc -all byte-identical, tests unchanged).
